[refactor] AnnotationHandlerMapping initialize 리팩토링
