--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/RequestMappedHandlerExecution.java
@@ -1,23 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.lang.reflect.Method;
-
-public class RequestMappedHandlerExecution extends HandlerExecution {
-
-    private final Object controllerInstance;
-    private final Method method;
-
-    public RequestMappedHandlerExecution(Object controllerInstance, Method method) {
-        this.controllerInstance = controllerInstance;
-        this.method = method;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        return (ModelAndView) method.invoke(controllerInstance, request, response);
-    }
-}
